backend/app/requirement.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
sqlmodel==0.0.8
python-dotenv==1.0.0
httpx==0.24.1
openai==1.0.0
pydantic==1.10.11
python-multipart==0.0.6
